+++ tests/unit/async_/fixtures/fake_connection.py
@@ -206,7 +206,7 @@ def async_scripted_connection_generator(async_fake_connection_generator):
                                 cb_args = default_cb_args
                             res = cb(*cb_args)
                             if cb_name == "on_failure":
-                                error = Neo4jError.hydrate(**cb_args[0])
+                                error = Neo4jError._hydrate_gql(**cb_args[0])
                             # suppress in case the callback is not async
                             with suppress(TypeError):
                                 await res