+++ tests/unit/common/test_exceptions.py
@@ -14,8 +14,16 @@
 # limitations under the License.
 
 
+from __future__ import annotations
+
+import contextlib
+import re
+import traceback
+
 import pytest
 
+import neo4j.exceptions
+from neo4j import PreviewWarning
 from neo4j._exceptions import (
     BoltError,
     BoltHandshakeError,
@@ -28,6 +36,7 @@ from neo4j.exceptions import (
     CLASSIFICATION_TRANSIENT,
     ClientError,
     DatabaseError,
+    GqlError,
     Neo4jError,
     ServiceUnavailable,
     TransientError,
@@ -148,8 +157,38 @@ def test_serviceunavailable_raised_from_bolt_protocol_error_with_explicit_style(
     assert e.value.__cause__ is error
 
 
+def _assert_default_gql_error_attrs_from_neo4j_error(error: GqlError) -> None:
+    with pytest.warns(PreviewWarning, match="GQLSTATUS"):
+        assert error.gql_status == "50N42"
+    if error.message:
+        with pytest.warns(PreviewWarning, match="GQLSTATUS"):
+            assert error.gql_status_description == (
+                "error: general processing exception - unexpected error. "
+                f"{error.message}"
+            )
+    else:
+        with pytest.warns(PreviewWarning, match="GQLSTATUS"):
+            assert error.gql_status_description == (
+                "error: general processing exception - unexpected error"
+            )
+    with pytest.warns(PreviewWarning, match="GQLSTATUS"):
+        assert (
+            error.gql_classification
+            == neo4j.exceptions.GqlErrorClassification.UNKNOWN
+        )
+    with pytest.warns(PreviewWarning, match="GQLSTATUS"):
+        assert error.gql_raw_classification is None
+    with pytest.warns(PreviewWarning, match="GQLSTATUS"):
+        assert error.diagnostic_record == {
+            "CURRENT_SCHEMA": "/",
+            "OPERATION": "",
+            "OPERATION_CODE": "0",
+        }
+    assert error.__cause__ is None
+
+
 def test_neo4jerror_hydrate_with_no_args():
-    error = Neo4jError.hydrate()
+    error = Neo4jError._hydrate_neo4j()
 
     assert isinstance(error, DatabaseError)
     assert error.classification == CLASSIFICATION_DATABASE
@@ -158,10 +197,13 @@ def test_neo4jerror_hydrate_with_no_args():
     assert error.metadata == {}
     assert error.message == "An unknown error occurred"
     assert error.code == "Neo.DatabaseError.General.UnknownError"
+    _assert_default_gql_error_attrs_from_neo4j_error(error)
 
 
-def test_neo4jerror_hydrate_with_message_and_code_rubish():
-    error = Neo4jError.hydrate(message="Test error message", code="ASDF_asdf")
+def test_neo4jerror_hydrate_with_message_and_code_rubbish():
+    error = Neo4jError._hydrate_neo4j(
+        message="Test error message", code="ASDF_asdf"
+    )
 
     assert isinstance(error, DatabaseError)
     assert error.classification == CLASSIFICATION_DATABASE
@@ -170,10 +212,11 @@ def test_neo4jerror_hydrate_with_message_and_code_rubish():
     assert error.metadata == {}
     assert error.message == "Test error message"
     assert error.code == "ASDF_asdf"
+    _assert_default_gql_error_attrs_from_neo4j_error(error)
 
 
 def test_neo4jerror_hydrate_with_message_and_code_database():
-    error = Neo4jError.hydrate(
+    error = Neo4jError._hydrate_neo4j(
         message="Test error message",
         code="Neo.DatabaseError.General.UnknownError",
     )
@@ -185,10 +228,11 @@ def test_neo4jerror_hydrate_with_message_and_code_database():
     assert error.metadata == {}
     assert error.message == "Test error message"
     assert error.code == "Neo.DatabaseError.General.UnknownError"
+    _assert_default_gql_error_attrs_from_neo4j_error(error)
 
 
 def test_neo4jerror_hydrate_with_message_and_code_transient():
-    error = Neo4jError.hydrate(
+    error = Neo4jError._hydrate_neo4j(
         message="Test error message",
         code="Neo.TransientError.General.TestError",
     )
@@ -200,10 +244,11 @@ def test_neo4jerror_hydrate_with_message_and_code_transient():
     assert error.metadata == {}
     assert error.message == "Test error message"
     assert error.code == f"Neo.{CLASSIFICATION_TRANSIENT}.General.TestError"
+    _assert_default_gql_error_attrs_from_neo4j_error(error)
 
 
 def test_neo4jerror_hydrate_with_message_and_code_client():
-    error = Neo4jError.hydrate(
+    error = Neo4jError._hydrate_neo4j(
         message="Test error message",
         code=f"Neo.{CLASSIFICATION_CLIENT}.General.TestError",
     )
@@ -215,6 +260,7 @@ def test_neo4jerror_hydrate_with_message_and_code_client():
     assert error.metadata == {}
     assert error.message == "Test error message"
     assert error.code == f"Neo.{CLASSIFICATION_CLIENT}.General.TestError"
+    _assert_default_gql_error_attrs_from_neo4j_error(error)
 
 
 @pytest.mark.parametrize(
@@ -252,9 +298,20 @@ def test_neo4jerror_hydrate_with_message_and_code_client():
         ),
     ),
 )
-def test_error_rewrite(code, expected_cls, expected_code):
+@pytest.mark.parametrize("mode", ("neo4j", "gql"))
+def test_error_rewrite(code, expected_cls, expected_code, mode):
     message = "Test error message"
-    error = Neo4jError.hydrate(message=message, code=code)
+    if mode == "neo4j":
+        error = Neo4jError._hydrate_neo4j(message=message, code=code)
+    elif mode == "gql":
+        error = Neo4jError._hydrate_gql(
+            gql_status="12345",
+            description="error: things - they hit the fan",
+            message=message,
+            neo4j_code=code,
+        )
+    else:
+        raise ValueError(f"Invalid mode {mode!r}")
 
     expected_retryable = expected_cls is TransientError
     assert error.__class__ is expected_cls
@@ -266,35 +323,55 @@ def test_error_rewrite(code, expected_cls, expected_code):
 
 
 @pytest.mark.parametrize(
-    ("code", "message", "expected_cls", "expected_str"),
+    ("code", "message", "expected_cls", "expected_str", "mode"),
     (
-        (
-            "Neo.ClientError.General.UnknownError",
-            "Test error message",
-            ClientError,
-            "{code: Neo.ClientError.General.UnknownError} "
-            "{message: Test error message}",
+        # values that behave the same in both modes
+        *(
+            (
+                *x,
+                mode,
+            )
+            for mode in ("neo4j", "gql")
+            for x in (
+                (
+                    "Neo.ClientError.General.UnknownError",
+                    "Test error message",
+                    ClientError,
+                    (
+                        "{code: Neo.ClientError.General.UnknownError} "
+                        "{message: Test error message}"
+                    ),
+                ),
+                (
+                    None,
+                    "Test error message",
+                    DatabaseError,
+                    (
+                        "{code: Neo.DatabaseError.General.UnknownError} "
+                        "{message: Test error message}"
+                    ),
+                ),
+                (
+                    "Neo.ClientError.General.UnknownError",
+                    None,
+                    ClientError,
+                    (
+                        "{code: Neo.ClientError.General.UnknownError} "
+                        "{message: An unknown error occurred}"
+                    ),
+                ),
+            )
         ),
-        (
-            None,
-            "Test error message",
-            DatabaseError,
-            "{code: Neo.DatabaseError.General.UnknownError} "
-            "{message: Test error message}",
-        ),
+        # neo4j error specific behavior
         (
             "",
             "Test error message",
             DatabaseError,
-            "{code: Neo.DatabaseError.General.UnknownError} "
-            "{message: Test error message}",
-        ),
-        (
-            "Neo.ClientError.General.UnknownError",
-            None,
-            ClientError,
-            "{code: Neo.ClientError.General.UnknownError} "
-            "{message: An unknown error occurred}",
+            (
+                "{code: Neo.DatabaseError.General.UnknownError} "
+                "{message: Test error message}"
+            ),
+            "neo4j",
         ),
         (
             "Neo.ClientError.General.UnknownError",
@@ -302,13 +379,39 @@ def test_error_rewrite(code, expected_cls, expected_code):
             ClientError,
             "{code: Neo.ClientError.General.UnknownError} "
             "{message: An unknown error occurred}",
+            "neo4j",
+        ),
+        # gql error specific behavior
+        (
+            "",
+            "Test error message",
+            DatabaseError,
+            "{code: } {message: Test error message}",
+            "gql",
+        ),
+        (
+            "Neo.ClientError.General.UnknownError",
+            "",
+            ClientError,
+            "{code: Neo.ClientError.General.UnknownError} {message: }",
+            "gql",
         ),
     ),
 )
 def test_neo4j_error_from_server_as_str(
-    code, message, expected_cls, expected_str
+    code, message, expected_cls, expected_str, mode
 ):
-    error = Neo4jError.hydrate(code=code, message=message)
+    if mode == "neo4j":
+        error = Neo4jError._hydrate_neo4j(code=code, message=message)
+    elif mode == "gql":
+        error = Neo4jError._hydrate_gql(
+            gql_status="12345",
+            description="error: things - they hit the fan",
+            neo4j_code=code,
+            message=message,
+        )
+    else:
+        raise ValueError(f"Invalid mode {mode!r}")
 
     assert type(error) is expected_cls
     assert str(error) == expected_str
@@ -320,3 +423,328 @@ def test_neo4j_error_from_code_as_str(cls):
 
     assert type(error) is cls
     assert str(error) == "Generated somewhere in the driver"
+
+
+def _make_test_gql_error(
+    identifier: str,
+    cause: GqlError | None = None,
+) -> GqlError:
+    error = GqlError(identifier)
+    error._init_gql(
+        gql_status=f"{identifier[:5].upper():<05}",
+        description=f"error: $h!t went down - {identifier}",
+        message=identifier,
+        cause=cause,
+    )
+    return error
+
+
+def _set_error_cause(exc, cause, method="set") -> None:
+    if method == "set":
+        exc.__cause__ = cause
+    elif method == "raise":
+        with contextlib.suppress(exc.__class__):
+            raise exc from cause
+    else:
+        raise ValueError(f"Invalid cause set method {method!r}")
+
+
+_CYCLIC_CAUSE_MARKER = object()
+
+
+def _assert_error_chain(
+    exc: BaseException,
+    expected: list[object],
+) -> None:
+    assert isinstance(exc, BaseException)
+
+    collection_root: BaseException | None = exc
+    actual_chain: list[object] = [exc]
+    actual_chain_ids = [id(exc)]
+    while collection_root is not None:
+        cause = getattr(collection_root, "__cause__", None)
+        if id(cause) in actual_chain_ids:
+            actual_chain.append(_CYCLIC_CAUSE_MARKER)
+            actual_chain_ids.append(id(_CYCLIC_CAUSE_MARKER))
+            break
+        actual_chain.append(cause)
+        actual_chain_ids.append(id(cause))
+        collection_root = cause
+
+    assert actual_chain_ids == list(map(id, expected))
+
+    expected_lines = [
+        str(exc)
+        for exc in expected
+        if exc is not None and exc is not _CYCLIC_CAUSE_MARKER
+    ]
+    expected_lines.reverse()
+    exc_fmt = traceback.format_exception(type(exc), exc, exc.__traceback__)
+    for line in exc_fmt:
+        if not expected_lines:
+            break
+        if expected_lines[0] in line:
+            expected_lines.pop(0)
+    if expected_lines:
+        traceback_fmt = "".join(exc_fmt)
+        pytest.fail(
+            f"Expected lines not found: {expected_lines} in traceback:\n"
+            f"{traceback_fmt}"
+        )
+
+
+def test_cause_chain_extension_no_cause() -> None:
+    root = _make_test_gql_error("root")
+
+    _assert_error_chain(root, [root, None])
+
+
+def test_cause_chain_extension_only_gql_cause() -> None:
+    root_cause = _make_test_gql_error("rootCause")
+    root = _make_test_gql_error("root", cause=root_cause)
+
+    _assert_error_chain(root, [root, root_cause, None])
+
+
+@pytest.mark.parametrize("local_cause_method", ("raise", "set"))
+def test_cause_chain_extension_only_local_cause(local_cause_method) -> None:
+    root_cause = ClientError("rootCause")
+    root = _make_test_gql_error("root")
+    _set_error_cause(root, root_cause, local_cause_method)
+
+    _assert_error_chain(root, [root, root_cause, None])
+
+
+@pytest.mark.parametrize("local_cause_method", ("raise", "set"))
+def test_cause_chain_extension_multiple_causes(local_cause_method) -> None:
+    root4_cause2 = _make_test_gql_error("r4c2")
+    root4_cause1 = _make_test_gql_error("r4c1", cause=root4_cause2)
+    root4 = _make_test_gql_error("root4", cause=root4_cause1)
+    root3 = ClientError("root3")
+    _set_error_cause(root3, root4, local_cause_method)
+    root2_cause3 = _make_test_gql_error("r2c3")
+    root2_cause2 = _make_test_gql_error("r2c2", cause=root2_cause3)
+    root2_cause1 = _make_test_gql_error("r2c1", cause=root2_cause2)
+    root2 = _make_test_gql_error("root2", cause=root2_cause1)
+    _set_error_cause(root2, root3, local_cause_method)
+    root1_cause2 = _make_test_gql_error("r1c2")
+    root1_cause1 = _make_test_gql_error("r1c1", cause=root1_cause2)
+    root1 = _make_test_gql_error("root1", cause=root1_cause1)
+    _set_error_cause(root1, root2, local_cause_method)
+
+    _assert_error_chain(
+        root1, [
+            root1, root1_cause1, root1_cause2,
+            root2, root2_cause1, root2_cause2, root2_cause3,
+            root3,
+            root4, root4_cause1, root4_cause2,
+            None,
+        ],
+    )  # fmt: skip
+
+
+@pytest.mark.parametrize("local_cause_method", ("raise", "set"))
+def test_cause_chain_extension_circular_local_causes(
+    local_cause_method,
+) -> None:
+    root6 = ClientError("root6")
+    root5 = _make_test_gql_error("root5")
+    _set_error_cause(root5, root6, local_cause_method)
+    root4_cause = _make_test_gql_error("r4c")
+    root4 = _make_test_gql_error("root4", cause=root4_cause)
+    _set_error_cause(root4, root5, local_cause_method)
+    root3 = ClientError("root3")
+    _set_error_cause(root3, root4, local_cause_method)
+    root2 = _make_test_gql_error("root2")
+    _set_error_cause(root2, root3, local_cause_method)
+    root1 = ClientError("root1")
+    _set_error_cause(root1, root2, local_cause_method)
+    _set_error_cause(root6, root1, local_cause_method)
+
+    _assert_error_chain(
+        root1,
+        [
+            root1,
+            root2,
+            root3,
+            root4,
+            root4_cause,
+            root5,
+            root6,
+            _CYCLIC_CAUSE_MARKER,
+        ],
+    )
+
+
+_DEFAULT_GQL_ERROR_ATTRIBUTES = {
+    "code": "Neo.DatabaseError.General.UnknownError",
+    "classification": "DatabaseError",
+    "category": "General",
+    "title": "UnknownError",
+    "message": "An unknown error occurred",
+    "gql_status": "50N42",
+    "gql_status_description": (
+        "error: general processing exception - unexpected error"
+    ),
+    "gql_classification": neo4j.exceptions.GqlErrorClassification.UNKNOWN,
+    "gql_raw_classification": None,
+    "diagnostic_record": {
+        "CURRENT_SCHEMA": "/",
+        "OPERATION": "",
+        "OPERATION_CODE": "0",
+    },
+    "__cause__": None,
+}
+
+
+@pytest.mark.parametrize(
+    ("metadata", "attributes"),
+    (
+        # all default values
+        (
+            {},
+            _DEFAULT_GQL_ERROR_ATTRIBUTES,
+        ),
+        # example from ADR
+        (
+            {
+                "gql_status": "01N00",
+                "message": "01EXAMPLE you have failed something",
+                "description": "client error - example error. Message",
+                "neo4j_code": "Neo.Example.Failure.Code",
+                "diagnostic_record": {
+                    "CURRENT_SCHEMA": "",
+                    "OPERATION": "",
+                    "OPERATION_CODE": "",
+                    "_classification": "CLIENT_ERROR",
+                    "_status_parameters": {},
+                },
+            },
+            {
+                "code": "Neo.Example.Failure.Code",
+                "classification": "Example",
+                "category": "Failure",
+                "title": "Code",
+                "message": "01EXAMPLE you have failed something",
+                "gql_status": "01N00",
+                "gql_status_description": (
+                    "client error - example error. Message"
+                ),
+                "gql_classification": (
+                    neo4j.exceptions.GqlErrorClassification.CLIENT_ERROR
+                ),
+                "gql_raw_classification": "CLIENT_ERROR",
+                "diagnostic_record": {
+                    "CURRENT_SCHEMA": "",
+                    "OPERATION": "",
+                    "OPERATION_CODE": "",
+                    "_classification": "CLIENT_ERROR",
+                    "_status_parameters": {},
+                },
+                "__cause__": None,
+            },
+        ),
+        # garbage diagnostic record
+        (
+            {
+                "diagnostic_record": {
+                    "CURRENT_SCHEMA": 1.5,
+                    "OPERATION": False,
+                    "_classification": ["whelp", None],
+                    "_🤡": "🎈",
+                    "foo": {"bar": "baz"},
+                },
+            },
+            {
+                **_DEFAULT_GQL_ERROR_ATTRIBUTES,
+                "gql_classification": (
+                    neo4j.exceptions.GqlErrorClassification.UNKNOWN
+                ),
+                "gql_raw_classification": None,
+                "diagnostic_record": {
+                    "CURRENT_SCHEMA": 1.5,
+                    "OPERATION": False,
+                    "_classification": ["whelp", None],
+                    "_🤡": "🎈",
+                    "foo": {"bar": "baz"},
+                },
+            },
+        ),
+        (
+            {
+                "diagnostic_record": {
+                    "_classification": "SOME_FUTURE_CLASSIFICATION",
+                },
+            },
+            {
+                **_DEFAULT_GQL_ERROR_ATTRIBUTES,
+                "gql_classification": (
+                    neo4j.exceptions.GqlErrorClassification.UNKNOWN
+                ),
+                "gql_raw_classification": "SOME_FUTURE_CLASSIFICATION",
+                "diagnostic_record": {
+                    "_classification": "SOME_FUTURE_CLASSIFICATION",
+                },
+            },
+        ),
+    ),
+)
+def test_gql_hydration(metadata, attributes):
+    # TODO: test causes
+    error = Neo4jError._hydrate_gql(**metadata)
+
+    preview_attrs = {
+        "gql_status",
+        "gql_status_description",
+        "gql_classification",
+        "gql_raw_classification",
+        "diagnostic_record",
+    }
+
+    for attr in (
+        "code",
+        "classification",
+        "category",
+        "title",
+        "message",
+        "gql_status",
+        "gql_status_description",
+        "gql_classification",
+        "gql_raw_classification",
+        "diagnostic_record",
+        "__cause__",
+    ):
+        expected_value = attributes[attr]
+        if attr in preview_attrs:
+            with pytest.warns(PreviewWarning, match="GQLSTATUS"):
+                actual_value = getattr(error, attr)
+        else:
+            actual_value = getattr(error, attr)
+        assert actual_value == expected_value
+
+
+@pytest.mark.parametrize(
+    "attr",
+    (
+        "code",
+        "classification",
+        "category",
+        "title",
+        "message",
+        "metadata",
+    ),
+)
+def test_deprecated_setter(attr):
+    obj = object()
+    error = Neo4jError()
+
+    with pytest.warns(
+        DeprecationWarning,
+        match=re.compile(
+            rf".*\baltering\b.*\b{attr}\b.*",
+            flags=re.IGNORECASE,
+        ),
+    ):
+        setattr(error, attr, obj)
+
+    assert getattr(error, attr) is obj